Pass UserID to Front End

+++ backend/routes/posts.js
@@ -49,7 +49,8 @@ router.post('', checkAuth, multer({storage: storage}).single('image'), (req, res
 });
 
 router.put(
-  '/:id', checkAuth,
+  '/:id',
+  checkAuth,
   multer({storage: storage}).single('image'),
   (req, res, next) => {
   let imagePath = req.body.imagePath;
@@ -63,9 +64,13 @@ router.put(
     content: req.body.content,
     imagePath: imagePath
   });
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    console.log(result);
-    res.status(200).json({ message: 'Update successful!' });
+  Post.updateOne({ _id: req.params.id, creator: req.userId.userId }, post).then(result => {
+    if(result.nModified > 0) {
+      res.status(200).json({ message: 'Update successful!' });
+    } else {
+      res.status(401).json({ message: 'Not authorized!' });
+    }
+    
   });
 });
 
@@ -104,9 +109,12 @@ router.get('/:id', (req, res, next) => {
 })
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({ message: 'Post Deleted' });
+  Post.deleteOne({ _id: req.params.id, creator: req.userId.userId }).then(result => {
+    if(result.n > 0) {
+      res.status(200).json({ message: 'Post Deleted' });
+    } else {
+      res.status(401).json({ message: 'Not authorized!' });
+    }
   }); 
 });
